// src/data/models/common/AiredStatus.ts
/**
 * Статус выхода произведения
 */
export enum AiredStatus {

    /**
     * Анонсировано
     */
    ANONS = "anons",

    /**
     * Выходит
     */
    ONGOING = "ongoing",

    /**
     * Вышло
     */
    RELEASED = "released",

    /**
     * Недавно вышедшее
     */
    LATEST = "latest",

    ///////////////////////////////////////////////////////////////////////////
    // Дополнительные статусы для манги
    ///////////////////////////////////////////////////////////////////////////

    /**
     * Паузы в выпуске манги
     */
    PAUSED = "paused",

    /**
     * Манга перестала выпускаться
     */
    DISCONTINUED = "discontinued",

    /**
     * Манга вышла
     */
    RELEASED_TWO = "finished_airing",

    /**
     * Манга выходит
     */
    ONGOING_TWO = "currently_airing",

    /**
     * Манга анонсирована
     */
    ANONS_TWO = "not_yet_aired",

    ///////////////////////////////////////////////////////////////////////////
    // Статусы, если пришла любая другая строка
    ///////////////////////////////////////////////////////////////////////////

    /**
     * Пока неизвестно
     */
    NONE = "none",

    /**
     * Неизвестный статус (если придёт что-то не из списка)
     */
    UNKNOWN = ''
}